41 add an option to disable new account activation via email in application properties

+++ rtc-communicator-microservice-main/src/main/java/project/rtc/registration/services/impl/RegistrationServiceImpl.java
@@ -12,10 +12,10 @@ import org.springframework.beans.factory.config.ConfigurableBeanFactory;
 import org.springframework.context.annotation.Scope;
 import org.springframework.stereotype.Service;
 
-import org.springframework.web.bind.MethodArgumentNotValidException;
 import project.rtc.authorization.basic_login.credentials.services.CredentialsService;
 import project.rtc.communicator.user.entities.User;
 import project.rtc.communicator.user.services.impl.UserServiceImpl;
+import project.rtc.infrastructure.exception.exceptions.InvalidTokenException;
 import project.rtc.infrastructure.utils.token.JwtTokenProvider;
 import project.rtc.registration.activateAccountToken.entities.ActivateAccountToken;
 import project.rtc.registration.activateAccountToken.services.impl.ActivateAccountTokenServiceImpl;
@@ -34,68 +34,43 @@ public class RegistrationServiceImpl implements RegistrationService {
 	private final UserServiceImpl userService;
 	@Value("${app.security.jwt.secret_key}")
 	private String jwtSecretKey;
+	@Value("${app.registration.security.active-account-via-email}")
+	private boolean activateAccountByEmailOption;
 
-	// User registration consists in creating new private credentials to which only the registrant has access
-	// and a public user account sent to friends during e.g. refreshing the friends list in the customer panel
-	public User registerAccount(RegistrationRequestDto dto) throws MethodArgumentNotValidException {
-		credentialsService.createCredentialsAndSaveInDatabase(dto);
-		User user = userService.create(dto.getNick(), dto.getEmail(), dto.getPicture());
-		// TODO: Uncomment after tests.
-		//sendActivateAccountLinkToEmail(registrationRequest.getEmail());
-		return user;
+	public User registerAccount(RegistrationRequestDto dto) throws MessagingException {
+		if(activateAccountByEmailOption){
+			credentialsService.create(dto, false);
+			sendActivateAccountLinkToEmail(dto.getEmail());
+		} else
+			credentialsService.create(dto, true);
+		return userService.create(dto.getNick(), dto.getEmail(), dto.getPicture());
 	}
 	
-	private boolean sendActivateAccountLinkToEmail(String email) {
+	private void sendActivateAccountLinkToEmail(String email) throws MessagingException {
 		ActivateAccountToken activateAccountToken = activateAccountTokenServiceImpl.assignNewTokenToAccount(email);
 		String link = "http://localhost:8080/app/registration/activate" + "?token=" + activateAccountToken.getToken();
-		String text = "Hello Marcin!\r\n"
+		mailSenderService.sendMessage(email, "Activate Link", prepareText(link));
+	}
+
+	public boolean activateAccount(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse)
+			throws InvalidTokenException {
+		String token = httpServletRequest.getParameter("token");
+		ActivateAccountToken activateAccountToken = activateAccountTokenServiceImpl.findByToken(token);
+		String email = JwtTokenProvider.getTokenSubject(jwtSecretKey, token);
+
+		// Check tokens are equal and token not expired.
+		if(JwtTokenProvider.getTokenExpiration(jwtSecretKey, token).before(new Date())
+				|| !activateAccountToken.getToken().equals(token))
+			throw new InvalidTokenException();
+
+		credentialsService.setAccountNonLocked(email, true);
+		activateAccountTokenServiceImpl.delete(email);
+		return true;
+	}
+
+	private String prepareText(String link){
+		return "Hello Marcin!\r\n"
 				+ "We received a request to create rtCommunicator account for you.\r\n"
 				+ "Enter the following link to activate account: " + link;
-		
-		try {
-			mailSenderService.sendMessage(email, "Acitave Link", text);
-			return true;
-		} catch (MessagingException e) {
-			e.printStackTrace();
-		}
-		return false;
 	}
-
-	public boolean activateAccount(HttpServletRequest httpServletRequest, HttpServletResponse httpServletResponse) {
-
-		String email;
-		ActivateAccountToken activateAccountToken;
-
-		String token = httpServletRequest.getParameter("token");
-		
-		try {
-			activateAccountToken = activateAccountTokenServiceImpl.findByToken(token);
-			email = JwtTokenProvider.getTokenSubject(jwtSecretKey, token);
-		} catch (Exception e) {
-			System.out.print(e.getMessage());
-			return false;
-		}
-
-		if(JwtTokenProvider.getTokenExpiration(jwtSecretKey, token).before(new Date()))
-			return false;
-
-			
-		if(activateAccountToken.getToken().equals(token)) {
-			credentialsService.setAccountNonLocked(email, true);
-			activateAccountTokenServiceImpl.delete(email);
-			try {	
-				httpServletResponse.sendRedirect(email);	
-			} catch (Exception e) {
-				System.out.print(e.getMessage());
-			}
-			return true;
-		}
-		
-		return false;
-	}
-	
-	public void activateAccount(String email) {
-		credentialsService.setAccountNonLocked(email, true);
-	}
-
 }
